Relocate files
